+++ CHANGELOG.md
@@ -19,7 +19,7 @@ export interface FormatterSettings extends FormattingOptions {
 - altered the signature of the `format(string, FormattingOptions)` function to `format(string, FormatterSettings)`, this is a non-breaking change as `FormatterSettings` extends `FormattingOptions` ([#62](https://github.com/rcjsuen/dockerfile-utils/issues/62))
 
 ### Fixed
-- do not validate variable substitutions if found in CMD and ENTRYPOINT ([#60](https://github.com/rcjsuen/dockerfile-utils/issues/60))
+- do not validate variable substitutions if found in CMD and ENTRYPOINT ([#89](https://github.com/rcjsuen/dockerfile-utils/issues/89))
 
 ## [0.2.0] - 2021-01-19
 ### Added